+++ CHANGELOG.rst
@@ -4,8 +4,17 @@ asfsmd release history
 asfsmd v1.1.0 (UNRELEASED)
 --------------------------
 
-* Add filter for polarization and additional calibration/noise/rfi downloads
-  (@scottstanie)
+* Now it is possible to customize the selection of files to be downloaded.
+  Beyond the manifest and the annotation files, now it is also possible to
+  download:
+
+  * calibration annotations
+  * noise annotations
+  * rfi annotations
+  * measurement files
+
+  Moreover now it is possible to select a specific beams or polarizations.
+  Patch developed by @scottstanie and @avalentino.
 
 
 asfsmd v1.0.0 (09/01/2022)